% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{caterpillar}
\alias{caterpillar}
\title{Caterpillar plot}
\usage{
caterpillar(results, ptm = FALSE, allPars = FALSE, byCond = FALSE)
}
\arguments{
\item{results}{The results list generated from the function
\code{\link{compCall}}}

\item{ptm}{A boolean variable denoting whether or not plots for ptm peptides
should be generated.}

\item{allPars}{A boolean variable indicating whether or not all parameters
should be plotted.  The default is false, resulting in a plot for only
parameters with 95% credible intervals that do not contain zero.}

\item{byCond}{A boolean parameter which determines if separate plots should
be made for each condition.  The default is FALSE which results in the
creation of a single plot.}
}
\description{
This function takes a list which contains a Stanfit object and summary
tables.  The output is a caterpillar plot of parameters with credible
intervals that do not contain 0.  Red line segments denote a 80% credible
interval, while the black tails show 95%.
}
