% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{checkVariance}
\alias{checkVariance}
\title{Variance plot}
\usage{
checkVariance(model)
}
\arguments{
\item{model}{A Stanmodel object generated from the function
\code{\link{compCall}}, which can be found in the third component of the
results object.}
}
\description{
This function shows the distribution of each parameter corresponding to an
experimental error.  Different variance components are modeled for each
tag/plex combination (along with different components for each PTM). If one
parameter is greatly different from the others it could be indicative of
problems in the experimental workflow.
}
