% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{compBayes}
\alias{compBayes}
\title{Fitting a compositional Stan model}
\usage{
compBayes(dat, approx = FALSE, resultsOnly = FALSE, pp = 0.95,
  nCores = 1, iter = 2000, nullSet = c(-0.2, 0.2), normalize = TRUE)
}
\arguments{
\item{dat}{The data to be analyzed.  This data must be formatted as shown in
the included sample data \code{\link{ptmDat}}.  Both the reference channel
 and the reference
conditions will be defined from the first column of data. If multiple runs
are to be analyzed at once then the dat object must be a list of
dataframes.  This is only recommended if biological or technical
replicates exist across multiple runs.  Otherwise, for computational
simplicity, each run should be analyzed separately.  For a single run, the
data should be formatted as a dataframe.  In each dataframe the first
two rows are used to determine which columns are technical replicates and
which are biological replicates.  For example, if two columns, possibly
from different plexes, both have the number '3' in the first row, then
they will be treated as technical replicates.  Likewise, columns that
share a number in the second row are treated as biological replicates. If
there are no biological replicates in the experiment then the second row
should be all zeroes.  If technical replicates are not also labeled as
biological replicates an error will be generated.}

\item{approx}{A boolean variable which specifies whether or not a full
Bayesian
or an approximate Bayesian model should be fit.  A value of TRUE
fits a Variational Bayes approximation.}

\item{resultsOnly}{A boolean variable that determines what information will
be returned after model fitting.  If resultsOnly is TRUE, then the
function will return a list containing two tables, one table for posterior
means and another with posterior standard deviations.  If resultsOnly is
FALSE then the results list will also contain an object of class Stanfit.}

\item{pp}{The percentage covariate level to be used for predicting relative
protein abundance.  By default this value is .95, so that sum signal to
noise will be adjusted towards the 95th percentile of observed values.}

\item{nCores}{determines the number of cores used for parallel processing
should be used.
For large datasets it is highly recommended that the package be used on a
computer capable of parallel processing.}

\item{iter}{Number of iterations for each chain}

\item{nullSet}{An interval representing unimportant changes.  This is
used to create the posterior probability that changes fall within the
unimportant interval.}

\item{normalize}{A boolean variable that determines whether or not
adjustments should be made under the assumption that the average protein
abundance is equivalent in each channel.  By default this value is true
which results in each row of the matrix being perturbed by the inverse of
the compositional mean.  Consequently the geometric means of each tag will
be equivalent.}
}
\description{
This function runs the main code for fitting an appropriate compositonal
proteomics model based on the
input data file.  It returns a list of size 3.  The first component of the
contains summary data on relative protein abundance.  The second component
summarizes PTM peptides and the third component contains the Stanfit object.
}
\details{
There are many details.  This will be filled out later.
}
