% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{compFit}
\alias{compFit}
\title{Fitting a compositional Stan model}
\usage{
compFit(dat, approx = TRUE)
}
\arguments{
\item{dat}{The data to be analyzed.  This data must be formatted as shown in
the included sample data.}

\item{approx}{A boolean variable which specifies whether or not a full Bayesian
or an approximate Bayesian model should be fit.  The default value of TRUE
fits a Variational Bayes approximation.}
}
\description{
This function fits an appropriate compositonal proteomics model based on the
input data file.  It returns a Stan modelfit object.
}
\details{
There are many details.  This will be filled out later.
}
