% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{compFreq}
\alias{compFreq}
\title{A frequentist version of the compositional model}
\usage{
compFreq(dat, normalize = TRUE, ssn = FALSE)
}
\arguments{
\item{dat}{The data to be analyzed.  This data must be formatted as shown in
the included sample data \code{\link{ptmDat}}.  Both the reference channel
and the reference
conditions will be defined from the first column of data. If multiple runs
are to be analyzed at once then the dat object must be a list of
dataframes.  This is only recommended if biological or technical
replicates exist across multiple runs.  Otherwise, for computational
simplicity, each run should be analyzed separately.  For a single run, the
data should be formatted as a dataframe.  In each dataframe the first
two rows are used to determine which columns are technical replicates and
which are biological replicates.  For example, if two columns, possibly
from different plexes, both have the number '3' in the first row, then
they will be treated as technical replicates.  Likewise, columns that
share a number in the second row are treated as biological replicates. If
there are no biological replicates in the experiment then the second row
should be all zeroes.  If technical replicates are not also labeled as
biological replicates an error will be generated.}

\item{normalize}{A boolean variable that determines whether or not
adjustments should be made under the assumption that the average protein
abundance is equivalent in each channel.  By default this value is true
which results in each row of the matrix being perturbed by the inverse of
the compositional mean.  Consequently the geometric means of each tag will
be equivalent.}
}
\description{
This function runs the main code for fitting an appropriate compositonal
proteomics model with a separate linear model for each protein.  It
returns a list of size 3.  The first component of the list
contains summary data on relative protein abundance.  The second component
summarizes PTM peptides and the third component contains asyptotic
standard errors generated from the models.
}
\details{
There are many details.  This will be filled out later.
}
