\name{compMS-package}
\alias{compMS-package}
\alias{compMS}
\docType{package}
\title{
\packageTitle{compMS}
}
\description{
\packageDescription{compMS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{compMS}
\packageIndices{compMS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{compMS}

Maintainer: \packageMaintainer{compMS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
