% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{contrastEst}
\alias{contrastEst}
\title{Changing the comparisons of interest}
\usage{
contrastEst(res, contrastMat = NULL, useCov = FALSE, stanKey = NULL)
}
\arguments{
\item{res}{The results object to be altered.}

\item{contrastMat}{A matrix specifying the contrasts of interest.
each row represents a contrast and the number of columns must match
the number of parameters.  If this is not specified then by default
tables for all possible pairwise comparisons will be generated.}
}
\description{
This function manipulates a Stanfit object to change the contrasts that
are estimated.
}
