% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{fitSimple}
\alias{fitSimple}
\title{Fitting the simple model}
\usage{
fitSimple(dat, approx, resultsOnly)
}
\description{
Function for specifically calling a model with no biological
replicates or condition groups.
}
