% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{propPlot}
\alias{propPlot}
\title{Proportion plot}
\usage{
propPlot(df, gene = NULL, condKey, protName = NULL)
}
\arguments{
\item{df}{A dataframe typically generated and stored in the second
component of a results list from compBayes.}

\item{gene}{A string that determines what data will be plotted.}

\item{condKey}{A matrix with two columns.  The first column contains
integers denoting the experimental condition for each biological
replicate.  The second contains integers that represent each
biological replicate.  This key should be taken from the
header information required to run the compBayes function.}
}
\description{
This function plots the proportion of ions belonging
to each biological replicate.  The first replicate
is typically a bridge channel and all repicates are
ordered and colored by experimental condition.
}
