% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testR.R
\name{test_function}
\alias{test_function}
\title{Testing this stuff out}
\usage{
test_function(y_)
}
\arguments{
\item{y_}{an input vector}
}
\value{
The posterior mean of beta
}
\description{
This code is just a way for me to learn roxygen and to test out
access to Stqn objects
}
\details{
We fit the model y = beta + epsilon.  Hopefully we will return
  the correct mean.
}
