% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longFunctions.R
\name{test_overall_effect}
\alias{test_overall_effect}
\title{This function runs the main code for fitting a model that looks at the
overall (single curve) effect through time.}
\usage{
test_overall_effect(tempDat, timeDegree = 2, fullTimes, useW = TRUE)
}
\arguments{
\item{tempDat}{The data to be analyzed}

\item{timeDegree}{either 1, 2 or 3 for linear, quadratic or cubic models}

\item{fullTimes}{A vector specifying all of the times found in the study}

\item{useW}{A boolean that determines whether or not the posterior variance of
each protein estimate should be used to weight the observations}
}
\description{
This function runs the main code for fitting a model that looks at the
overall (single curve) effect through time.
}
