% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longFunctions.R
\name{testInteract}
\alias{testInteract}
\title{This function runs the main code for fitting a model that compares more
than one curve through time.}
\usage{
testInteract(tempDat, timeDegree = 2, fullTimes, fullCats, useW = TRUE,
  refCat = NULL, groupByGene = FALSE, randEffect = 0)
}
\arguments{
\item{tempDat}{The data to be analyzed}

\item{timeDegree}{either 1, 2 or 3 for linear, quadratic or cubic models}

\item{fullTimes}{A vector specifying all of the times found in the study}

\item{fullCats}{A vector specifying all of the categories found in the study}

\item{useW}{A boolean that determines whether or not the posterior variance of
each protein estimate should be used to weight the observations}

\item{refCat}{A string that specifies the reference category}

\item{groupByGene}{A boolean that specifies whether timepoints are grouped
by genes or proteins}

\item{randEffect}{takes 0 for no random effects. 1 for a random intercept
and 2 for a random slope and intercepts model}
}
\description{
This function runs the main code for fitting a model that compares more
than one curve through time.
}
