% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{toSimplex}
\alias{toSimplex}
\title{A transformation from log ratio space into proportions}
\usage{
toSimplex(RES)
}
\arguments{
\item{RES}{Results from the compBayes function.}
}
\description{
This function transforms all estimated log ratios into proportions
}
\details{
Estimates from compBayes are done after using the additive log ratio
  transformation.  This function applies the inverse transformation to the
  full sampling chains and recomputes summary statistics.  Returns a data
  frame with estimates, variance and intervals in terms of proportions.
}
