% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS-package.R
\docType{package}
\name{compMS-package}
\alias{compMS-package}
\alias{compMS}
\title{Bayesian Compositional Proteomics Modeling via Stan}
\description{
Bayesian Compositional Proteomics Modeling via Stan
}
