% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS-package.R
\docType{package}
\name{compMS-package}
\alias{compMS-package}
\alias{compMS}
\title{Bayesian Compositional Proteomics Modeling via Stan}
\description{
The \pkg{compMS} package enables users to fit compostional proteomics
models for isobaric tag mass spectrometry experiments.  The package is
designed to make these models accesible to people without a background
in Bayesian statistics or compositional data analysis.  This is done by
requiring users to follow a very specific format for data entry.  Once
the data is formatted properly everything else is handled internally.
The required fields and headers are explained in the documentation for the
sample data \code{\link{sampleDat}}.

Once data has been correctly formatted and loaded into your workspace, you can
create a Stan model with the function \code{\link{compBayes}}.
}
\section{Model Details}{

This is where I will explain the model and the options for use:
\describe{
No description yet.  For now see the supplement from the manuscript
https://pubs.acs.org/doi/10.1021/acs.jproteome.7b00699
}
}

