% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{precisionPlot}
\alias{precisionPlot}
\title{Precision plot}
\usage{
precisionPlot(summary)
}
\arguments{
\item{summary}{A results table generated from the function
\code{\link{compCall}}, which will be found in either the first (for
protein summaries) or second (for ptm's) component of the result list
object.}
}
\description{
This function takes a results summary from the \code{\link{compCall}}
function.
The output is a plot of showing posterior means on the
x-axis and precision on the y-axis, where precision defined as the inverse
of the posterior variance.
}
